feat: move things around for better performance and add some loaders
