+++ packages/widget-playground/src/components/Widget/WidgetViewContainer.tsx
@@ -6,7 +6,7 @@ import { ExternalWalletProvider } from '../../providers';
 import {
   useConfig,
   useEditToolsActions,
-  useEditToolsValues,
+  useDrawerToolValues,
 } from '../../store';
 import { ToggleDrawerButton } from './ToggleDrawerButton';
 import {
@@ -25,7 +25,7 @@ export function WidgetViewContainer({
   toggleDrawer,
 }: WidgetViewContainerProps) {
   const { config } = useConfig();
-  const { isDrawerOpen, drawerWidth } = useEditToolsValues();
+  const { isDrawerOpen, drawerWidth } = useDrawerToolValues();
   const { setDrawerOpen } = useEditToolsActions();
 
   const isWalletManagementExternal = !!config?.walletConfig;
