+++ packages/widget-playground/src/store/editTools/index.ts
@@ -1,4 +1,6 @@
 export * from './EditToolsProvider';
-export * from './useEditToolsValues';
+export * from './useDrawerToolValues';
 export * from './useEditToolsActions';
+export * from './useFontToolValues';
+export * from './useCodeToolValues';
 export * from './constants';
