+++ packages/widget-playground/src/store/editTools/useCodeToolValues.ts
@@ -1,0 +1,13 @@
+import { shallow } from 'zustand/shallow';
+import { useEditToolsStore } from './EditToolsProvider';
+
+export const useCodeToolsValues = () => {
+  const [codeControlTab] = useEditToolsStore(
+    (store) => [store.codeControl.openTab],
+    shallow,
+  );
+
+  return {
+    codeControlTab,
+  };
+};
