+++ packages/widget-playground/src/store/editTools/useDrawerToolValues.ts
@@ -1,0 +1,28 @@
+import { shallow } from 'zustand/shallow';
+import { useEditToolsStore } from './EditToolsProvider';
+import { defaultDrawerWidth } from './constants';
+
+export const useDrawerToolValues = () => {
+  const [isDrawerOpen, codeDrawerWidth, visibleControls, codeControlTab] =
+    useEditToolsStore(
+      (store) => [
+        store.drawer.open,
+        store.drawer.codeDrawerWidth,
+        store.drawer.visibleControls,
+        store.codeControl.openTab,
+        store.fontControl.selectedFont,
+      ],
+      shallow,
+    );
+
+  const drawerWidth =
+    visibleControls === 'code' && codeControlTab === 'config'
+      ? codeDrawerWidth
+      : defaultDrawerWidth;
+
+  return {
+    isDrawerOpen,
+    visibleControls,
+    drawerWidth,
+  };
+};
