+++ packages/widget-playground/src/store/editTools/useFontToolValues.ts
@@ -1,0 +1,13 @@
+import { shallow } from 'zustand/shallow';
+import { useEditToolsStore } from './EditToolsProvider';
+
+export const useFontToolsValues = () => {
+  const [selectedFont] = useEditToolsStore(
+    (store) => [store.fontControl.selectedFont],
+    shallow,
+  );
+
+  return {
+    selectedFont,
+  };
+};
