+++ packages/widget-playground/src/store/widgetConfig/index.ts
@@ -2,3 +2,4 @@ export * from './WidgetConfigProvider';
 export * from './useConfigActions';
 export * from './useConfigValues';
 export * from './useConfig';
+export * from './utils/getWhitelistedConfig';
